udpate photoswipe info dialog
