+++ src/components/PhotoSwipe/styledComponents/Legend.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Legend = styled.span`
+    font-size: 20px;
+    color: #ddd;
+    display: inline;
+`;
