+++ src/components/PhotoSwipe/styledComponents/LegendContainer.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const LegendContainer = styled.div`
+    display: flex;
+    justify-content: space-between;
+`;
