+++ src/components/PhotoSwipe/styledComponents/Pre.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Pre = styled.pre`
+    color: #aaa;
+    padding: 7px 15px;
+`;
