+++ src/components/PhotoSwipe/styledComponents/SmallLoadingSpinner.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import EnteSpinner from 'components/EnteSpinner';
+
+export const SmallLoadingSpinner = () => (
+    <EnteSpinner
+        style={{
+            width: '20px',
+            height: '20px',
+        }}
+    />
+);
