+++ src/constants/photoswipe/index.ts
@@ -1,0 +1,7 @@
+export const defaultLivePhotoDefaultOptions = {
+    click: () => {},
+    hide: () => {},
+    show: () => {},
+    loading: false,
+    visible: false,
+};
