+++ src/utils/strings/englishConstants.tsx
@@ -463,21 +463,21 @@ const englishConstants = {
     VIDEO_PLAYBACK_FAILED: 'video format not supported',
     VIDEO_PLAYBACK_FAILED_DOWNLOAD_INSTEAD:
         'this video cannot be played on your browser',
-    METADATA: 'metadata',
-    INFO: 'information',
-    FILE_ID: 'file id',
-    FILE_NAME: 'file name',
-    CREATION_TIME: 'creation time',
-    UPDATED_ON: 'updated on',
-    LOCATION: 'location',
+    METADATA: 'Metadata',
+    INFO: 'Info',
+    FILE_ID: 'File id',
+    FILE_NAME: 'File name',
+    CREATION_TIME: 'Creation time',
+    UPDATED_ON: 'Updated on',
+    LOCATION: 'Location',
     SHOW_MAP: 'show on map',
-    EXIF: 'exif',
-    DEVICE: 'device',
-    IMAGE_SIZE: 'image size',
-    FLASH: 'flash',
-    FOCAL_LENGTH: 'focal length',
-    APERTURE: 'aperture',
-    ISO: 'iso',
+    EXIF: 'Exif',
+    DEVICE: 'Device',
+    IMAGE_SIZE: 'Image size',
+    FLASH: 'Flash',
+    FOCAL_LENGTH: 'Focal length',
+    APERTURE: 'Aperture',
+    ISO: 'ISO',
     SHOW_ALL: 'show all',
     LOGIN_TO_UPLOAD_FILES: (count: number) =>
         count === 1
